Update the core data structure
With the new ability of bit sets as a container for the indices of set
bits, now we barely have the need to use separate vector to store
dimension subsets any more.  Now basically all dimension subsets are
implemented as bits, except inside the broken summation iterator, where
an index vector is still internally used to translate internal summation
index to the actual global one.